+++ AddOns/Baganator/Skins/ElvUI.lua
@@ -147,9 +147,13 @@ local skinners = {
   end,
   CheckBox = function(frame)
     S:HandleCheckBox(frame)
+    frame:SetSize(20, 20)
   end,
   Slider = function(frame)
-    S:HandleStepSlider(frame)
+    frame:SetHeight(12)
+    S:HandleSliderFrame(frame.Slider)
+    S:HandleNextPrevButton(frame.Back)
+    S:HandleNextPrevButton(frame.Forward)
   end,
   InsetFrame = function(frame)
     if frame.NineSlice then
@@ -188,6 +192,7 @@ end
 
 local function SetConstants()
   addonTable.Constants.ButtonFrameOffset = 0
+  addonTable.Constants.ButtonFrameOffsetTop = 0
 end
 
 local function LoadSkin()
@@ -220,5 +225,5 @@ if addonTable.Skins.IsAddOnLoading("ElvUI") then
       rightText = addonTable.Locales.RELOAD_REQUIRED,
       default = false,
     },
-  }, true, addonTable.Skins.PlayerLoginLoadingTrigger)
+  }, true)
 end