+++ AddOns/Baganator/CHANGELOG.md
@@ -1,6 +1,7 @@
 # Baganator
 
-## [820](https://github.com/TheMouseNest/Baganator/tree/820) (2026-08-13)
-[Full Changelog](https://github.com/TheMouseNest/Baganator/compare/819...820) 
+## [829](https://github.com/TheMouseNest/Baganator/tree/829) (2026-09-21)
+[Full Changelog](https://github.com/TheMouseNest/Baganator/compare/828...829) 
 
-- Workaround Blizzard bootstrap change instantly closing guild bank  
+- Forever: Fix reagent bags not being separated  
+- Forever: Fix sorting into reagent bags